\name{lenna}
\alias{lenna}
\docType{data}
\title{Lenna image represented as a three dimensional array}
\description{
\code{lenna} is a 461*480*3 numeric array containing the RGB channels values of each pixel of the 461*480 sized image
}
\usage{data("lenna")}
\format{
  The format is:
 num [1:461, 1:480, 1:3] 0.788 0.776 0.776 0.792 0.796 ...
}
\source{Playboy... Taper lena jpeg dans google...
}
\examples{
data(lenna)
## maybe str(lenna) ; plot(lenna) ...
}
\keyword{datasets}
