% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACPJPEG.R
\name{pcajpeg}
\alias{pcajpeg}
\title{PCA on the RGB channels of a JPEG file}
\usage{
pcajpeg(file = NULL)
}
\arguments{
\item{file}{Character string specifying the path to the JPEG file to analyse}
}
\value{
Produces two JPEG files named img1.jpeg and img2.jpeg.
img1.jpeg is obtained from file by conserving only the first principal component.
img2.jpeg is obtained from file by conserving only the two first principal component(s).
}
\description{
JPEG images are converted into a n*3 matrix, where n stands for the number of pixels in the image and
the columns are the Red, Green and Blue (RGB) channels of the pixels. Then, a PCA is performed.
Outputs are two JPEG images obtained by projecting the original one to the (two) principal component(s).
}
\examples{
data(lenna)
jpeg::writeJPEG(lenna, target='lenna.jpeg')
pcajpeg(file="lenna.jpeg")
}
